+++ sshd-common/src/main/java/org/apache/sshd/client/config/keys/ClientIdentityLoaderHolder.java
@@ -17,13 +17,18 @@
  * under the License.
  */
 
-package org.apache.sshd.common.config.keys;
+package org.apache.sshd.client.config.keys;
 
 /**
  * TODO Add javadoc
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public interface FilePasswordProviderManager extends FilePasswordProviderHolder {
-    void setFilePasswordProvider(FilePasswordProvider provider);
+@FunctionalInterface
+public interface ClientIdentityLoaderHolder {
+    /**
+     * @return The {@link ClientIdentityLoader} to use in order to load client
+     * key pair identities - never {@code null}
+     */
+    ClientIdentityLoader getClientIdentityLoader();
 }